Added ownership check for client management

+++ Systems/API/ZooExpoOrg.Api/Controllers/Clients/ClientController.cs
@@ -5,9 +5,9 @@ using ZooExpoOrg.Services.Logger;
 using ZooExpoOrg.Services.Clients;
 using Microsoft.IdentityModel.Tokens;
 using ZooExpoOrg.Common.Exceptions;
-using Azure.Core;
 using Microsoft.AspNetCore.Authorization;
 using ZooExpoOrg.Common.Security;
+using ZooExpoOrg.Services.RightVerifier;
 
 namespace ZooExpoOrg.Api.Controllers.Clients;
 
@@ -20,12 +20,19 @@ public class ClientController : Controller
     private readonly IAppLogger logger;
     private readonly IClientService clientService;
     private readonly IMapper mapper;
+    private readonly IRightVerifierService rightVerifier;
 
-    public ClientController(IAppLogger logger, IClientService clientService, IMapper mapper)
+    public ClientController(
+        IAppLogger logger, 
+        IClientService clientService, 
+        IMapper mapper,
+        IRightVerifierService rightVerifier
+        )
     {
         this.logger = logger;
         this.clientService = clientService;
         this.mapper = mapper;
+        this.rightVerifier = rightVerifier;
     }
 
     [HttpGet("")]
@@ -56,6 +63,13 @@ public class ClientController : Controller
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfCreateClient(jwtToken, model.UserId)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             var result = await clientService.Create(mapper.Map<CreateClientModel>(model));
 
             return Ok(mapper.Map<PresintationClientModel>(result));
@@ -72,6 +86,13 @@ public class ClientController : Controller
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagClient(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await clientService.Update(id, mapper.Map<UpdateClientModel>(model));
 
             return Ok();
@@ -88,6 +109,13 @@ public class ClientController : Controller
     {
         try
         {
+            string jwtToken = Request.Headers["Authorization"].ToString().Replace("Bearer ", "");
+
+            if (!(await rightVerifier.VerifRightsOfManagClient(jwtToken, id)))
+            {
+                return BadRequest("Access denied.");
+            }
+
             await clientService.Delete(id);
 
             return Ok();
